+++ tests/Enums/AnnotateFixtures/EnumWithMultipleLineComments.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace BenSampo\Enum\Tests\Enums;
+namespace BenSampo\Enum\Tests\Enums\Annotate;
 
 use BenSampo\Enum\Enum;