+++ tests/Enums/AnnotateOriginals/EnumWithSingleLineComment.php
@@ -1,15 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace BenSampo\Enum\Tests\Enums;
+namespace BenSampo\Enum\Tests\Enums\Annotate;
 
 use BenSampo\Enum\Enum;
 
 /**
  * This is a test enum with single line comment.
  *
- * @method static static A()
- * @method static static B()
- * @method static static C()
  * @extends Enum<self::*>
  */
 final class EnumWithSingleLineComment extends Enum